backend/faiss_requirements.txt
fastapi
pydantic
faiss
numpy
pymongo-4.14.1
